setting up the foundations

+++ app/src/main/java/com/example/adam/myusefulllocations/Data/CurrentLocation.java
@@ -2,7 +2,7 @@ package com.example.adam.myusefulllocations.Data;
 
 public interface CurrentLocation {
 
-    public void currentLocation (long lat, long lon, String currentAddress);
+    public void currentLocation (double lat, double lon, String currentAddress);
 
 
 }
